+++ apm-sniffer/apm-sdk-plugin/sharding-sphere-4.x-RC3-plugin/src/main/java/org/apache/skywalking/apm/plugin/shardingsphere/v4rc3/define/JDBCRootInvokeInstrumentation.java
@@ -16,7 +16,7 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.shardingsphere.v4.define;
+package org.apache.skywalking.apm.plugin.shardingsphere.v4rc3.define;
 
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.matcher.ElementMatcher;
@@ -29,15 +29,15 @@ import org.apache.skywalking.apm.agent.core.plugin.match.NameMatch;
 import static net.bytebuddy.matcher.ElementMatchers.named;
 
 /**
- * {@link ParseInstrumentation} presents that skywalking intercepts {@link org.apache.shardingsphere.core.route.router.sharding.ShardingRouter}.
+ * {@link JDBCRootInvokeInstrumentation} presents that skywalking intercepts {@link org.apache.shardingsphere.shardingjdbc.executor.AbstractStatementExecutor}.
  *
  * @author zhangyonglun
  */
-public class ParseInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
+public class JDBCRootInvokeInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
     
-    private static final String ENHANCE_CLASS = "org.apache.shardingsphere.core.route.router.sharding.ShardingRouter";
+    private static final String ENHANCE_CLASS = "org.apache.shardingsphere.shardingjdbc.executor.AbstractStatementExecutor";
     
-    private static final String EXECUTE_INTERCEPTOR_CLASS = "org.apache.skywalking.apm.plugin.shardingsphere.v4.ParseInterceptor";
+    private static final String JDBC_ROOT_INVOKE_INTERCEPTOR_CLASS = "org.apache.skywalking.apm.plugin.shardingsphere.v4rc3.JDBCRootInvokeInterceptor";
     
     @Override
     public InstanceMethodsInterceptPoint[] getInstanceMethodsInterceptPoints() {
@@ -45,12 +45,12 @@ public class ParseInstrumentation extends ClassInstanceMethodsEnhancePluginDefin
             new InstanceMethodsInterceptPoint() {
                 @Override
                 public ElementMatcher<MethodDescription> getMethodsMatcher() {
-                    return named("parse");
+                    return named("executeCallback");
                 }
                 
                 @Override
                 public String getMethodsInterceptor() {
-                    return EXECUTE_INTERCEPTOR_CLASS;
+                    return JDBC_ROOT_INVOKE_INTERCEPTOR_CLASS;
                 }
                 
                 @Override
